# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/detective": {
            "get": {
                "description": "Endpoint ini mengambil seluruh data detektif dengan dukungan paginasi. Cocok untuk monitoring personel investigasi. Asumsi: data konsisten, paginasi default 10 item per halaman.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Detectives"
                ],
                "summary": "Mendapatkan seluruh data detektif",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nomor halaman, default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah item per halaman, default 10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar detektif beserta metadata paginasi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/detective/create": {
            "post": {
                "description": "Endpoint ini melakukan proses pembuatan detektif baru, termasuk validasi data, transformasi input, dan penanganan error. Proses ini penting untuk memastikan integritas data personel investigasi. Asumsi: data valid, badge_number unik. Batasan: hanya admin dapat akses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Detectives"
                ],
                "summary": "Membuat entitas detektif baru",
                "parameters": [
                    {
                        "description": "Payload detektif dalam format JSON.\n\nField 'investigation_style' hanya dapat bernilai salah satu dari: \n- 'Evidence-Based Investigation': Pendekatan investigasi yang menitikberatkan pada pengumpulan, analisis, dan interpretasi bukti fisik secara sistematis. Metode ini mengedepankan objektivitas dan validitas ilmiah dalam setiap tahapan investigasi. \n- 'Interview-Based Investigation': Strategi investigasi yang berfokus pada teknik wawancara mendalam terhadap saksi, korban, dan tersangka. Pendekatan ini menuntut keterampilan komunikasi interpersonal dan analisis psikologis tingkat tinggi. \n- 'Undercover Investigation': Metode investigasi di mana detektif menyamar untuk memperoleh informasi yang tidak dapat diakses secara terbuka. Pendekatan ini sangat berisiko namun efektif dalam mengungkap kejahatan terorganisir. \n- 'Follow The Money Investigation': Investigasi yang memprioritaskan pelacakan aliran dana untuk mengidentifikasi motif, pelaku, dan jaringan kejahatan. Pendekatan ini sangat relevan dalam kasus korupsi dan kejahatan finansial. \n- 'Report-Based Investigation': Pendekatan yang mengandalkan analisis laporan, dokumen, dan data administratif sebagai sumber utama informasi. Cocok untuk kasus yang membutuhkan audit forensik dan penelusuran administratif secara mendalam.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/detective_request.Detective_Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Detektif berhasil dibuat; response berisi data detektif baru",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Permintaan tidak valid; data input salah atau format tidak sesuai",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/detective/delete/{id}": {
            "delete": {
                "description": "Endpoint ini menghapus data detektif berdasarkan ID. Proses ini tidak dapat dibatalkan dan berdampak pada assignment kasus terkait. Asumsi: ID valid dan eksis.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Detectives"
                ],
                "summary": "Menghapus data detektif",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID detektif yang akan dihapus",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus detektif",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/detective/update/{id}": {
            "put": {
                "description": "Endpoint ini memperbarui data detektif berdasarkan ID, mendukung perubahan seluruh atribut utama. Validasi dan error handling dilakukan secara ketat. Asumsi: ID valid, badge_number tetap unik.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Detectives"
                ],
                "summary": "Memperbarui data detektif",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID detektif yang akan diperbarui",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload detektif dalam format JSON.\n\nField 'investigation_style' hanya dapat bernilai salah satu dari: \n- 'Evidence-Based Investigation': Pendekatan investigasi yang menitikberatkan pada pengumpulan, analisis, dan interpretasi bukti fisik secara sistematis. Metode ini mengedepankan objektivitas dan validitas ilmiah dalam setiap tahapan investigasi. \n- 'Interview-Based Investigation': Strategi investigasi yang berfokus pada teknik wawancara mendalam terhadap saksi, korban, dan tersangka. Pendekatan ini menuntut keterampilan komunikasi interpersonal dan analisis psikologis tingkat tinggi. \n- 'Undercover Investigation': Metode investigasi di mana detektif menyamar untuk memperoleh informasi yang tidak dapat diakses secara terbuka. Pendekatan ini sangat berisiko namun efektif dalam mengungkap kejahatan terorganisir. \n- 'Follow The Money Investigation': Investigasi yang memprioritaskan pelacakan aliran dana untuk mengidentifikasi motif, pelaku, dan jaringan kejahatan. Pendekatan ini sangat relevan dalam kasus korupsi dan kejahatan finansial. \n- 'Report-Based Investigation': Pendekatan yang mengandalkan analisis laporan, dokumen, dan data administratif sebagai sumber utama informasi. Cocok untuk kasus yang membutuhkan audit forensik dan penelusuran administratif secara mendalam.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/detective_request.Detective_Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data detektif",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Permintaan tidak valid; data input salah",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/detective/{id}": {
            "get": {
                "description": "Endpoint ini mengambil data detektif spesifik berdasarkan UUID. Cocok untuk detail view atau proses assignment kasus. Asumsi: ID valid dan eksis.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Detectives"
                ],
                "summary": "Mendapatkan detektif berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID detektif yang dicari",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data detektif",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan sistem internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "detective_request.Detective_Request": {
            "type": "object",
            "required": [
                "investigation_style"
            ],
            "properties": {
                "badge_number": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "investigation_style": {
                    "type": "string",
                    "enum": [
                        "Evidence-Based Investigation",
                        "Interview-Based Investigation",
                        "Undercover Investigation",
                        "Follow The Money Investigation",
                        "Report-Based Investigation"
                    ]
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "station": {
                    "type": "string"
                }
            }
        },
...
    }
}